Changes to be committed: 	modified:   .gitignore 	deleted:    db/init.sql 	modified:   docker-compose.yaml 	modified:   fastApiEL/src/db/models.py

+++ fastApiEL/src/db/models.py
@@ -1,6 +1,6 @@
 import uuid
 
-from sqlalchemy import INTEGER, Column, VARCHAR
+from sqlalchemy import BIGINT, Column, VARCHAR
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import declarative_base
 
@@ -11,7 +11,7 @@ class User(Base):
     __tablename__ = "users"
 
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    id_tg = Column(INTEGER, nullable=False, unique=True)
-    education_id = Column(INTEGER, nullable=True)
-    group_id = Column(INTEGER, nullable=True)
+    id_tg = Column(BIGINT, nullable=False, unique=True)
+    education_id = Column(BIGINT, nullable=True)
+    group_id = Column(BIGINT, nullable=True)
     jwt_token = Column(VARCHAR, nullable=True)
